requirements.txt
streamlit
pandas
numpy
plotly
yfinance
fpdf2
requests
